refactor: rework copyDir

+++ 04-copy-directory/index.js
@@ -1,28 +1,35 @@
 const fsPromises = require('fs').promises;
 const path = require('path');
 
-async function copyDir() {
-  const filesFolderPath = path.join(__dirname, 'files');
-  const filesCopyFolderPath = path.join(__dirname, 'files-copy');
+const filesFolderPath = path.join(__dirname, 'files');
+const filesCopyFolderPath = path.join(__dirname, 'files-copy');
 
-  await fsPromises.mkdir(filesCopyFolderPath, { recursive: true });
-  await clearFolder(filesCopyFolderPath);
-  await copyFiles(filesFolderPath, filesCopyFolderPath);
+async function copyDir(sourcePath, destinationPath) {
+  await fsPromises.mkdir(destinationPath, { recursive: true });
+  await clearFolder(destinationPath);
+  await copyDirContent(sourcePath, destinationPath);
 }
 
 async function clearFolder(folderPath) {
   const copyFiles = await fsPromises.readdir(folderPath);
   for (const fileName of copyFiles) {
-    await fsPromises.unlink(path.resolve(folderPath, fileName));
+    const folderChildPath = path.resolve(folderPath, fileName);
+    await fsPromises.rm(folderChildPath, { recursive: true });
   }
 }
-async function copyFiles(sourcePath, destinationPath) {
+
+async function copyDirContent(sourcePath, destinationPath) {
   const files = await fsPromises.readdir(sourcePath);
   for (const fileName of files) {
     const source = path.join(sourcePath.toString(), fileName);
     const destination = path.join(destinationPath.toString(), fileName);
-    await fsPromises.copyFile(source, destination);
+    const stat = await fsPromises.stat(source);
+    if (stat.isFile()) {
+      await fsPromises.copyFile(source, destination);
+    } else {
+      await copyDir(source, destination);
+    }
   }
 }
 
-copyDir().catch(console.error);
+copyDir(filesFolderPath, filesCopyFolderPath).catch(console.error);
